add base problems
