+++ algorithm/arrayRela/27_removeElement_test.go
@@ -1,0 +1,3 @@
+package arrayRela
+
+// see ../twoPointer/27_removeElement_test.go
