+++ algorithm/arrayRela/704_search_test.go
@@ -1,0 +1,3 @@
+package arrayRela
+
+// see ../binarySearch/704_search_test.go
