+++ algorithm/twoPointer/27_removeElement_test.go
@@ -1,5 +1,10 @@
 package twoPointer
 
+import (
+	"github.com/stretchr/testify/assert"
+	"testing"
+)
+
 func removeElement(nums []int, val int) int {
 	slow, fast := 0, 0
 	for ; fast < len(nums); fast++ {
@@ -9,4 +14,26 @@ func removeElement(nums []int, val int) int {
 		}
 	}
 	return slow
+}
+
+func TestRemoveElement(t *testing.T) {
+	tests := []struct {
+		nums   []int
+		val    int
+		output int
+	}{
+		{
+			nums:   []int{3, 2, 2, 3},
+			val:    3,
+			output: 2,
+		},
+		{
+			nums:   []int{0, 1, 2, 2, 3, 0, 4, 2},
+			val:    2,
+			output: 5,
+		},
+	}
+	for _, test := range tests {
+		assert.Equal(t, test.output, removeElement(test.nums, test.val))
+	}
 }
